UPSTREAM: 26161: kubectl: move printObjectSpecificMessage in factory
