+++ Godeps/_workspace/src/k8s.io/kubernetes/pkg/kubectl/cmd/replace.go
@@ -151,7 +151,7 @@ func RunReplace(f *cmdutil.Factory, out io.Writer, cmd *cobra.Command, args []st
 		}
 
 		info.Refresh(obj, true)
-		printObjectSpecificMessage(obj, out)
+		f.PrintObjectSpecificMessage(obj, out)
 		cmdutil.PrintSuccess(mapper, shortOutput, out, info.Mapping.Resource, info.Name, "replaced")
 		return nil
 	})
@@ -244,7 +244,7 @@ func forceReplace(f *cmdutil.Factory, out io.Writer, cmd *cobra.Command, args []
 
 		count++
 		info.Refresh(obj, true)
-		printObjectSpecificMessage(obj, out)
+		f.PrintObjectSpecificMessage(obj, out)
 		cmdutil.PrintSuccess(mapper, shortOutput, out, info.Mapping.Resource, info.Name, "replaced")
 		return nil
 	})
